pymor.operators.basic.ProjectedOperator -> pymor.operators.constructions.ProjectedOperator
