+++ src/pymor/operators/basic.py
@@ -121,119 +121,3 @@ class LinearComplexifiedListVectorArrayOperatorBase(ListVectorArrayOperatorBase)
         else:
             imag_part = None
         return self.range.complexified_vector_type(real_part, imag_part)
-
-
-class ProjectedOperator(Operator):
-    """Generic |Operator| representing the projection of an |Operator| to a subspace.
-
-    This operator is implemented as the concatenation of the linear combination with
-    `source_basis`, application of the original `operator` and projection onto
-    `range_basis`. As such, this operator can be used to obtain a reduced basis
-    projection of any given |Operator|. However, no offline/online decomposition is
-    performed, so this operator is mainly useful for testing before implementing
-    offline/online decomposition for a specific application.
-
-    This operator is instantiated in :func:`pymor.algorithms.projection.project`
-    as a default implementation for parametric or nonlinear operators.
-
-    Parameters
-    ----------
-    operator
-        The |Operator| to project.
-    range_basis
-        See :func:`pymor.algorithms.projection.project`.
-    source_basis
-        See :func:`pymor.algorithms.projection.project`.
-    product
-        See :func:`pymor.algorithms.projection.project`.
-    solver_options
-        The |solver_options| for the projected operator.
-    """
-
-    linear = False
-
-    def __init__(self, operator, range_basis, source_basis, product=None, solver_options=None):
-        assert isinstance(operator, Operator)
-        assert source_basis is None or source_basis in operator.source
-        assert range_basis is None or range_basis in operator.range
-        assert (product is None
-                or (isinstance(product, Operator)
-                    and range_basis is not None
-                    and operator.range == product.source
-                    and product.range == product.source))
-        if source_basis is not None:
-            source_basis = source_basis.copy()
-        if range_basis is not None:
-            range_basis = range_basis.copy()
-        self.__auto_init(locals())
-        self.build_parameter_type(operator)
-        self.source = NumpyVectorSpace(len(source_basis)) if source_basis is not None else operator.source
-        self.range = NumpyVectorSpace(len(range_basis)) if range_basis is not None else operator.range
-        self.linear = operator.linear
-
-    @property
-    def H(self):
-        if self.product:
-            return super().H
-        else:
-            options = {'inverse': self.solver_options.get('inverse_adjoint'),
-                       'inverse_adjoint': self.solver_options.get('inverse')} if self.solver_options else None
-            return ProjectedOperator(self.operator.H, self.source_basis, self.range_basis, solver_options=options)
-
-    def apply(self, U, mu=None):
-        mu = self.parse_parameter(mu)
-        if self.source_basis is None:
-            if self.range_basis is None:
-                return self.operator.apply(U, mu=mu)
-            elif self.product is None:
-                return self.range.make_array(self.operator.apply2(self.range_basis, U, mu=mu).T)
-            else:
-                V = self.operator.apply(U, mu=mu)
-                return self.range.make_array(self.product.apply2(V, self.range_basis))
-        else:
-            UU = self.source_basis.lincomb(U.to_numpy())
-            if self.range_basis is None:
-                return self.operator.apply(UU, mu=mu)
-            elif self.product is None:
-                return self.range.make_array(self.operator.apply2(self.range_basis, UU, mu=mu).T)
-            else:
-                V = self.operator.apply(UU, mu=mu)
-                return self.range.make_array(self.product.apply2(V, self.range_basis))
-
-    def jacobian(self, U, mu=None):
-        if self.linear:
-            return self.assemble(mu)
-        assert len(U) == 1
-        mu = self.parse_parameter(mu)
-        if self.source_basis is None:
-            J = self.operator.jacobian(U, mu=mu)
-        else:
-            J = self.operator.jacobian(self.source_basis.lincomb(U.to_numpy()), mu=mu)
-        from pymor.algorithms.projection import project
-        pop = project(J, range_basis=self.range_basis, source_basis=self.source_basis,
-                      product=self.product)
-        if self.solver_options:
-            options = self.solver_options.get('jacobian')
-            if options:
-                pop = pop.with_(solver_options=options)
-        return pop
-
-    def assemble(self, mu=None):
-        op = self.operator.assemble(mu=mu)
-        if op == self.operator:  # avoid infinite recursion in apply_inverse default impl
-            return self
-        from pymor.algorithms.projection import project
-        pop = project(op, range_basis=self.range_basis, source_basis=self.source_basis,
-                      product=self.product)
-        if self.solver_options:
-            pop = pop.with_(solver_options=self.solver_options)
-        return pop
-
-    def apply_adjoint(self, V, mu=None):
-        assert V in self.range
-        if self.range_basis is not None:
-            V = self.range_basis.lincomb(V.to_numpy())
-        U = self.operator.apply_adjoint(V, mu)
-        if self.source_basis is not None:
-            U = self.source.make_array(U.dot(self.source_basis))
-        return U
